# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SignalFlow API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SaveLeadRequest {
  name: string;
  title?: string;
  profile_url: string;
  /** LinkedIn About section text, used for AI draft generation */
  about_text?: string;
}

export interface Lead {
  id: number;
  name: string;
  title?: string | null;
  profile_url: string;
  ai_draft?: string | null;
  created_at: string;
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}