# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SignalFlow API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: leads
    description: Lead management operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /save-lead:
    post:
      operationId: saveLead
      tags: [leads]
      summary: Save a LinkedIn lead
      description: Saves a scraped LinkedIn profile as a lead. Requires API key authentication.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveLeadRequest"
      responses:
        "201":
          description: Lead saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: Get all leads
      description: Returns all saved leads
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
  /leads/{id}:
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      description: Deletes a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SaveLeadRequest:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        profile_url:
          type: string
        about_text:
          type: string
          description: LinkedIn About section text, used for AI draft generation
      required:
        - name
        - profile_url
    Lead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
          nullable: true
        profile_url:
          type: string
        ai_draft:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - profile_url
        - created_at
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error