# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SignalFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Lead {
  id: number;
  name: string;
  title?: string | null;
  profile_url: string;
  ai_draft?: string | null;
  created_at: Date;
}