# lib/api-zod/src/generated/types/saveLeadRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SignalFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SaveLeadRequest {
  name: string;
  title?: string;
  profile_url: string;
  /** LinkedIn About section text, used for AI draft generation */
  about_text?: string;
}